refactor(e2e): address PR review round 5 feedback
- Split 1090-line e2e_test.go into focused files: setup_test.go,
  fixture_test.go, configure_test.go, helpers_test.go, radarr_test.go,
  sonarr_test.go
- Change run() to return error instead of int; TestMain uses os.Exit(1)
  on error, skipping the explicit os.Exit(run(m)) pattern
- Add composePull() step with 20-minute timeout before composeUp(),
  separate from the service-health wait
- Remove forceRemoveDir: containers now run as user 1000:1000 (host user)
  so volume files are owned by the current user
- Replace syntheticHash (FNV-1a) with extractBTIH, which reads the btih
  directly from the magnet URI's xt=urn:btih: parameter
- Switch Sonarr test series from The Lone Ranger (TVDB 72059, 26-min
  episodes) to Colonel Bleep (TVDB 254376, 6-min episodes, public domain).
  BBB at ~9:56 is 166% of the 6-min episode runtime, clearing Sonarr's
  50% sample check, so no special MKV fixture is needed
- Use BBB MP4 directly as the qbt stub fixture for both tests; remove
  ensureShortMKVFixture
- Use make build target instead of direct go build in startProcesses
- Add ffprobe assertions to both tests: container=matroska, codec=hevc,
  duration>300s
- findMKV returns string path and fs.SkipAll after first match
- Fix single-letter loop variable names throughout (s→service, f→file/entry,
  h→hash, k/v→name/cat, ep→episode)
- Add rule to CLAUDE.md prohibiting single-letter loop iteration variables
- Add comments to docker-compose.yml explaining max_connections=1000 and
  user: "1000:1000"
- Remove accidental blank line from pkg/ffmpeg/stream_video.go
