+++ CLAUDE.md
@@ -70,6 +70,7 @@ All required tools (`go`, `golangci-lint`, etc.) are provided by `flake.nix`. If
 - Separate test cases that require fundamentally different setup into their own test functions.
 - Integration tests that require external services (e.g. a running Hatchet server) belong in files with a `//go:build integration` build tag. Skip with `t.Skip(...)` if required env vars are absent. Run via `make test-integration`.
 - Always use `t.Context()` (not `context.Background()`) when a test needs a context. It is automatically cancelled when the test finishes, preventing resource leaks.
+- Loop iteration variable names must be the singular form of the collection name, not single-letter abbreviations. For example, use `service` when ranging over `services`, `entry` when ranging over `entries`. Single-letter names like `s`, `v`, `f`, `k` are not allowed.
 
 ## Quality Rules
